+++ src/main/java/custom/components/CreateIndexComp.java
@@ -208,7 +208,7 @@ public class CreateIndexComp {
                 .failCount(failCount)
                 .truncated(truncated)
                 .totalCostTime(totalCost)
-                .rps(totalCost > 0 ? successCount / totalCost : 0)
+                .rps(totalCost > 0 ? successCount / (double) totalCost : 0d)
                 .avg(MathUtil.calculateAverage(costTimeTotal))
                 .tp99(MathUtil.calculateTP99(costTimeTotal, 0.99f))
                 .tp98(MathUtil.calculateTP99(costTimeTotal, 0.98f))
